pf-sort - remove module, update directive to export a function
