--- src/common/components/sort-directive.js
@@ -1,226 +1,0 @@
-/**
- * @ngdoc directive
- * @name patternfly.sort.directive:pfSort
- *
- * @description
- *   Directive for a  sort component
- *   <br><br>
- *
- * @param {object} config configuration settings for the sort:<br/>
- * <ul style='list-style-type: none'>
- * <li>.fields          - (Array) List of sortable fields containing:
- * <ul style='list-style-type: none'>
- * <li>.id          - (String) Unique Id for the sort field
- * <li>.title       - (String) The title to display for the sort field
- * <li>.sortType    - (String) The sort type, 'alpha' or 'numeric'
- * </ul>
- * <li>.currentField   - (Object) Currently selected field
- * <li>.isAscending - (boolean) Current sort direction is ascending. True for ascending, False for descending
- * <li>.onSortChange - ( function(sortId, sortDirection ) Function to call when the current sort params change
- * </ul>
- *
- * @example
-<example module="patternfly.sort">
-  <file name="index.html">
-    <div ng-controller="ViewCtrl" class="row example-container">
-      <div class="col-md-12">
-        <div pf-sort id="exampleSort" config="sortConfig"></div>
-      </div>
-      <hr class="col-md-12">
-      <div class="col-md-12">
-        <label class="events-label">Items: </label>
-      </div>
-      <div class="col-md-12">
-        <div ng-repeat="item in items" class="col-md-12 cfme-row-column">
-          <div class="row">
-            <div class="col-md-3">
-              <span>{{item.name}}</span>
-            </div>
-            <div class="col-md-3">
-              <span>{{item.count}}</span>
-            </div>
-            <div class="col-md-3">
-              <span>{{item.description}}</span>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </file>
-
-  <file name="script.js">
-    angular.module('patternfly.sort').controller('ViewCtrl', ['$scope',
-      function ($scope) {
-        $scope.items = [
-          {
-            name: "Item 7",
-            count: 432,
-            description: 'Very nice item'
-          },
-          {
-            name: "Item 6",
-            count: 22,
-            description: 'It lasts forever'
-          },
-          {
-            name: "Item 3",
-            count: 632,
-            description: 'Good stuff cheap'
-          },
-          {
-            name: "Item 2",
-            count: 12,
-            description: 'Fantastic'
-          },
-          {
-            name: "Item 9",
-            count: 99,
-            description: 'It does alright'
-          },
-          {
-            name: "Item 4",
-            count: 442,
-            description: 'Horrible'
-          },
-          {
-            name: "Item 1",
-            count: 42,
-            description: 'Most excellent'
-          },
-          {
-            name: "Item 8",
-            count: 2,
-            description: 'Get it while it lasts'
-          },
-          {
-            name: "Item 5",
-            count: 321,
-            description: 'Beautiful style'
-          }
-        ];
-
-        var compareFn = function(item1, item2) {
-          var compValue = 0;
-          if ($scope.sortConfig.currentField.id === 'name') {
-            compValue = item1.name.localeCompare(item2.name);
-          } else if ($scope.sortConfig.currentField.id === 'count') {
-              compValue = item1.count - item2.count;
-          } else if ($scope.sortConfig.currentField.id === 'description') {
-            compValue = item1.description.localeCompare(item2.description);
-          }
-
-          if (!$scope.sortConfig.isAscending) {
-            compValue = compValue * -1;
-          }
-
-          return compValue;
-        };
-
-        var sortChange = function (sortId, isAscending) {
-          $scope.items.sort(compareFn);
-        };
-
-        $scope.sortConfig = {
-          fields: [
-            {
-              id: 'name',
-              title:  'Name',
-              sortType: 'alpha'
-            },
-            {
-              id: 'count',
-              title:  'Count',
-              sortType: 'numeric'
-            },
-            {
-              id: 'description',
-              title:  'Description',
-              sortType: 'alpha'
-            }
-          ],
-          onSortChange: sortChange
-        };
-      }
-    ]);
-  </file>
-</example>
- */
-angular.module('patternfly.sort').directive('pfSort', function () {
-  'use strict';
-  return {
-    restrict: 'A',
-    scope: {
-      config: '='
-    },
-    templateUrl: 'sort/sort.html',
-    controller: function ($scope) {
-
-      $scope.setupConfig = function () {
-        var updated = false;
-
-        if ($scope.config.fields === undefined) {
-          $scope.config.fields = [];
-        }
-
-        if ($scope.config.fields.length > 0) {
-          if ($scope.config.currentField === undefined) {
-            $scope.config.currentField = $scope.config.fields[0];
-            updated = true;
-          }
-          if ($scope.config.isAscending === undefined) {
-            $scope.config.isAscending = true;
-            updated = true;
-          }
-        }
-
-        if (updated === true && $scope.config.onSortChange) {
-          $scope.config.onSortChange($scope.config.currentField, $scope.config.isAscending);
-        }
-      };
-
-      $scope.selectField = function (field) {
-        $scope.config.currentField = field;
-
-        if ($scope.config.onSortChange) {
-          $scope.config.onSortChange($scope.config.currentField, $scope.config.isAscending);
-        }
-      };
-
-      $scope.changeDirection = function () {
-        $scope.config.isAscending = !$scope.config.isAscending;
-
-        if ($scope.config.onSortChange) {
-          $scope.config.onSortChange($scope.config.currentField, $scope.config.isAscending);
-        }
-      };
-
-      $scope.getSortIconClass = function () {
-        var iconClass;
-
-        if ($scope.config.currentField.sortType === 'numeric') {
-          if ($scope.config.isAscending) {
-            iconClass = 'fa fa-sort-numeric-asc';
-          } else {
-            iconClass = 'fa fa-sort-numeric-desc';
-          }
-        } else {
-          if ($scope.config.isAscending) {
-            iconClass = 'fa fa-sort-alpha-asc';
-          } else {
-            iconClass = 'fa fa-sort-alpha-desc';
-          }
-        }
-
-        return iconClass;
-      };
-
-      $scope.setupConfig();
-    },
-
-    link: function (scope, element, attrs) {
-      scope.$watch('config', function () {
-        scope.setupConfig();
-      }, true);
-    }
-  };
-});
