--- src/common/components/sort.module.js
@@ -1,8 +1,0 @@
-/**
- * @name  patternfly card
- *
- * @description
- *   Sort module for patternfly.
- *
- */
-angular.module('patternfly.sort', ['ui.bootstrap']);
